requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-nested-admin==4.0.2
django-tinymce==4.0.0
djangorestframework==3.15.1
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
